backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.30.0
psycopg[binary]>=3.2.10
alembic==1.14.1
celery[redis]==5.6.2
redis==5.2.1
gitpython==3.1.46
cryptography==44.0.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.20
PyGithub==2.6.0
python-gitlab==5.6.0
azure-devops==7.1.0b4
httpx==0.28.1
langchain>=0.3
langchain-community>=0.3
litellm>=1.60
sse-starlette>=2.0
